+++ src/main/java/it/uiip/digitalgarage/roboadvice/businesslogic/model/dto/StrategyDTO.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 @Getter @Setter @AllArgsConstructor @NoArgsConstructor @Builder @ToString
 public class StrategyDTO  {
 
-    private Integer assetClassId;
+    private Long assetClassId;
 
     private BigDecimal percentage;
