+++ src/main/java/it/uiip/digitalgarage/roboadvice/persistence/repository/AssetClassRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.uiip.digitalgarage.roboadvice.persistence.model.AssetClass;
 
-public interface AssetClassRepository extends JpaRepository<AssetClass, Integer> {
+public interface AssetClassRepository extends JpaRepository<AssetClass, Long> {
 
 }
