+++ src/main/java/it/uiip/digitalgarage/roboadvice/persistence/repository/AssetRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AssetRepository extends JpaRepository<Asset, Integer> {
+public interface AssetRepository extends JpaRepository<Asset, Long> {
 
     List<Asset> findByAssetClass(AssetClass assetClass);
